+++ CHANGELOG.md
@@ -52,6 +52,8 @@ with crypto fix
 * [#95](https://github.com/babylonlabs-io/covenant-emulator/pull/95) removed local signer option
 as the covenant emulator should only connect to a remote signer
 * [#96](https://github.com/babylonlabs-io/covenant-emulator/pull/96) add pagination to `queryDelegationsWithStatus`
+* [#98](https://github.com/babylonlabs-io/covenant-emulator/pull/98) add submission of covenant
+signatures as one msg per transaction if batch of transactions fails
 * [#99](https://github.com/babylonlabs-io/covenant-emulator/pull/99) add more metrics
 covenant-signer
 * [#100](https://github.com/babylonlabs-io/covenant-emulator/pull/100) bump babylon to v1.0.0-rc.4